requirements.txt
altgraph==0.17.4
colorama==0.4.6
decorator==5.2.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
keyboard==0.13.5
moviepy==2.2.1
numpy==2.3.2
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
pillow_heif==1.1.0
proglog==0.1.12
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pyperclip==1.9.0
python-dotenv==1.1.1
pywin32-ctypes==0.2.3
setuptools==80.9.0
tqdm==4.67.1
